Address Phase 5 optional dependency review feedback
